update telegram bot id

+++ modules/core/src/main/java/org/mycontroller/standalone/settings/TelegramBotSettings.java
@@ -57,7 +57,7 @@ public class TelegramBotSettings {
         }
         return TelegramBotSettings.builder()
                 .token(token)
-                .id(McUtils.getInteger(getValue(SKEY_ID)))
+                .id(McUtils.getLong(getValue(SKEY_ID)))
                 .isBot(McUtils.getBoolean(getValue(SKEY_IS_BOT)))
                 .firstName(getValue(SKEY_FIRST_NAME))
                 .username(getValue(SKEY_USERNAME)).build();
